refactor(svelte): split interval query pure unit tests by export
Move shared IntervalQueryScene builder to query-fixtures and split the
652-line query.test.ts into resolve-parts, pixels/domains, and
selection/scene suites co-located with query.ts exports.
